+++ yaralyzer/util/argument_parser.py
@@ -210,6 +210,9 @@ export.add_argument('-sfx', '--file-suffix',
                     metavar='SUFFIX',
                     help='optional string to use as the suffix for exported files of any kind')
 
+tuning.add_argument('--no-timestamps', action='store_true',
+                    help="do not append file creation timestamps to exported filenames")
+
 
 # Debugging
 debug = parser.add_argument_group(
@@ -310,13 +313,17 @@ def parse_arguments(args: Optional[Namespace] = None):
         log_current_config()
         log_argparse_result(YaralyzerConfig.args, 'with_env_vars')
 
-    # print(f"yaralyzer parse_args() complete")
     return args
 
 
-def get_export_basepath(args: Namespace, yaralyzer: Yaralyzer) -> str:
+def get_export_basepath(args: Namespace, yaralyzer: Yaralyzer) -> Path:
     """Get the basepath (directory + filename without extension) for exported files."""
     args.output_basename  = f"{args.file_prefix}{yaralyzer._filename_string()}"  # noqa: E221
     args.output_basename += f"__maxdecode{YaralyzerConfig.args.max_decode_length}"
     args.output_basename += args.file_suffix
-    return path.join(args.output_dir, args.output_basename + f"__at_{args.invoked_at_str}")
+    file_basename = args.output_basename
+
+    if not args.no_timestamps:
+        file_basename += f"__at_{args.invoked_at_str}"
+
+    return args.output_dir.joinpath(file_basename)